fix: 금액 input 컴포넌트 분리
